r13711: * Correctly handle acb_info/acct_flags as uint32 not as uint16. * Fix a couple of related parsing issues. * in the info3 reply in a samlogon, return the ACB-flags (instead of   returning zero)
Guenther

+++ source/include/libmsrpc.h
@@ -1761,7 +1761,7 @@ struct SamEnumUsers {
       POLICY_HND *dom_hnd;
 
       /**Enumerate users with specific ACB. If 0, all users will be enumerated*/
-      uint16 acb_mask;
+      uint32 acb_mask;
    } in;
 
    struct {
